requirements.txt
bcrypt==3.1.4
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Jinja2==2.10
Pillow==5.3.0
pycparser==2.18
six==1.11.0
SQLAlchemy==1.2.7
WTForms==2.1
tensorflow==1.14
keras==2.2.5
numpy==1.17.2
pymongo==3.11.4 

